Overflow bug fixed.
Former-commit-id: 78fae0b4c330b26858060698c531d86cd8efd25f
Former-commit-id: d0362c314d6c5047691e1346aff652bfa622b651

+++ App/src/DataFile/gdf2.h
@@ -50,8 +50,8 @@ private:
 	FILE* file;
 	double samplingFrequency;
 	uint64_t samplesRecorded;
-	int startOfData;
-	int startOfEventTable;
+	size_t startOfData;
+	size_t startOfEventTable;
 	bool isLittleEndian;
 	double* scale;
 	int dataTypeSize;
